--- client/src/components/feedDetail/presentational/CommentItem.js
@@ -1,38 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const style = {};
-
-style.CommentItem = styled.div`
-  display: flex;
-  font-size: 14px;
-  padding: 0 20px 20px 20px;
-`;
-
-style.AuthorProfileImg = styled.img`
-  width: 30px;
-  height: 30px;
-  border-radius: 15px;
-  overflow: hidden;
-  margin-right: 14px;
-`;
-
-style.CommentContent = styled.div`
-  width: 200px;
-  word-break: break-all;
-`;
-
-const CommentItem = ({ comment }) => {
-  return (
-    <style.CommentItem>
-      <style.AuthorProfileImg src={comment.author.profileImg} />
-      <style.CommentContent>
-        <b>{comment.author.userName}</b>
-        &nbsp;
-        {comment.content}
-      </style.CommentContent>
-    </style.CommentItem>
-  );
-};
-
-export default CommentItem;
