+++ client/src/components/home/container/HomeContainer.js
@@ -1,9 +1,9 @@
-import React, { useState, useContext } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
 import styled from 'styled-components';
 import Side from '@home/presentational/Side';
 import Contents from '@home/presentational/Contents';
+import pathURL from '@constants/path';
 import UserContext from '@context/user';
-import HomeFeedAPI from '@api/HomeFeedAPI';
 
 const style = {};
 
@@ -14,19 +14,32 @@ style.HomeContainer = styled.div`
   margin: 0 auto;
   margin-top: 30px;
 `;
-
 const HomeContainer = () => {
   const [data, setData] = useState([]);
   const [loading, setLoading] = useState(false);
-  const [getMore, setGetMore] = useState('noId');
   const { login } = useContext(UserContext);
 
-  HomeFeedAPI(data, setData, loading, setLoading, login, getMore);
-
+  const getData = () => {
+    const url = pathURL.IP + pathURL.API_HOME_FEED + login.userName;
+    const option = {
+      mode: 'cors',
+      method: 'GET',
+    };
+    async function fetchUrl() {
+      const response = await fetch(url, option);
+      const json = await response.json();
+      setData(json);
+      setLoading(true);
+    }
+    useEffect(() => {
+      fetchUrl();
+    }, []);
+  };
+  getData();
   if (loading) {
     return (
       <style.HomeContainer>
-        <Contents data={data} setGetMore={setGetMore} />
+        <Contents data={data} />
         <Side />
       </style.HomeContainer>
     );
