+++ client/src/components/home/presentational/Comment.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 import UserName from '@home/presentational/UserName';
-import { Link } from 'react-router-dom';
 
 const style = {};
 
@@ -17,20 +16,11 @@ style.Content = styled.div`
   text-overflow: ellipsis;
 `;
 
-style.Link = styled(Link)`
-  text-decoration: none;
-`;
-
 const Comment = (input) => {
   const { author, content } = input;
-  const { userName } = author;
-  const href = `/profile?userName=${userName}`;
-
   return (
     <style.Comment>
-      <style.Link to={href}>
-        <UserName>{userName}</UserName>
-      </style.Link>
+      <UserName>{author.userName}</UserName>
       <style.Content>{content}</style.Content>
     </style.Comment>
   );
