+++ client/src/components/home/presentational/Feed.js
@@ -7,22 +7,10 @@ const style = {};
 style.Feed = styled.div``;
 
 const Feed = (input) => {
-  const { data: datas, setGetMore } = input;
-
+  const { data: datas } = input;
   return (
     <style.Feed>
-      {datas.map((data, index) => {
-        const isLastItem = datas.length === index + 1;
-        if (isLastItem) {
-          return (
-            <FeedItem
-              key={data._id}
-              data={data}
-              isLastItem
-              setGetMore={setGetMore}
-            />
-          );
-        }
+      {datas.map((data) => {
         return <FeedItem key={data._id} data={data} />;
       })}
     </style.Feed>
