+++ client/src/components/home/presentational/UserName.js
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 
 const UserName = styled.div`
   width: auto;
+  margin: auto 0;
   margin-left: 16px;
   font-weight: 600;
   font-size: 14px;
