+++ client/src/components/profile/container/ProfileContainer.js
@@ -1,10 +1,9 @@
-import React, { useState, useRef, useContext } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
 import styled from 'styled-components';
 import ProfileInfo from '@profile/presentational/ProfileInfo';
 import FeedList from '@feedExplore/presentational/FeedList';
+import pathURL from '@constants/path';
 import UserContext from '@context/user';
-import ModalContext from '@context/modal';
-import ProfileFeedAPI from '@api/ProfileFeedAPI';
 
 const style = {};
 
@@ -12,35 +11,31 @@ style.ProfileContainer = styled.div``;
 
 const ProfileContainer = () => {
   const { login } = useContext(UserContext);
-  let userName = window.location.search.split('=')[1];
-  const [preUserName, setPreUserName] = useState(userName);
-  if (userName !== preUserName) setPreUserName(userName);
-  if (!userName) userName = login.userName;
+  const { userName } = login;
   const [data, setData] = useState([]);
   const [loading, setLoading] = useState(false);
-  const [feeds, setFeeds] = useState([]);
-  const { modalActive } = useContext(ModalContext);
-  const [getMore, setGetMore] = useState('noId');
-  const isMounted = useRef(false);
-
-  ProfileFeedAPI(
-    setData,
-    loading,
-    setLoading,
-    getMore,
-    userName,
-    preUserName,
-    feeds,
-    setFeeds,
-    isMounted,
-    modalActive,
-  );
-
-  if (loading && data.feeds) {
+  const getData = () => {
+    const url = pathURL.IP + pathURL.API_PROFILE + userName;
+    const option = {
+      mode: 'cors',
+      method: 'GET',
+    };
+    async function fetchUrl() {
+      const response = await fetch(url, option);
+      const json = await response.json();
+      setData(json);
+      setLoading(true);
+    }
+    useEffect(() => {
+      fetchUrl();
+    }, []);
+  };
+  getData();
+  if (loading) {
     return (
       <style.ProfileContainer>
-        <ProfileInfo login={login} data={data} userInfo={data.userInfo} />
-        <FeedList datas={feeds} setGetMore={setGetMore} />
+        <ProfileInfo data={data} />
+        <FeedList datas={data.feeds.reverse()} />
       </style.ProfileContainer>
     );
   }
