+++ client/src/components/style/GlobalStyle.js
@@ -4,7 +4,6 @@ const GlobalStyle = createGlobalStyle`
 	body {
 		padding: 0;
     margin: 0;
-		background-color: ${(props) => props.theme.color.background};
 	}
 `;
