--- client/src/constants/value.js
@@ -1,8 +1,0 @@
-const initLogin = {
-  jwt: '',
-  name: '',
-  userName: '',
-  profileImg: '',
-};
-
-export default initLogin;
