+++ client/webpack.config.js
@@ -35,7 +35,7 @@ module.exports = {
   plugins: [
     new HtmlWebpackPlugin({
       template: './src/index.html',
-      favicon: './src/public/favicon_duck.ico',
+      favicon: './src/public/favicon.ico',
     }),
     new CleanWebpackPlugin(),
   ],
@@ -51,12 +51,6 @@ module.exports = {
       '@constants': path.resolve(__dirname, 'src/constants/'),
       '@context': path.resolve(__dirname, 'src/context/'),
       '@login': path.resolve(__dirname, 'src/components/login/'),
-      '@feedDetail': path.resolve(__dirname, 'src/components/feedDetail/'),
-      '@public': path.resolve(__dirname, 'src/public/'),
-      '@hooks': path.resolve(__dirname, 'src/hooks/'),
-      '@api': path.resolve(__dirname, 'src/api/'),
-      '@utils': path.resolve(__dirname, 'src/utils/'),
-      '@search': path.resolve(__dirname, 'src/components/search/'),
     },
   },
   devtool: 'inline-source-map',
