+++ server/src/routes/index.ts
@@ -5,7 +5,6 @@ import imageRouter from './image.router';
 import loginRouter from './login.router';
 import userRouter from './user.router';
 import commentRouter from './comment.router';
-import searchRouter from './search.router';
 
 const router = express.Router();
 
@@ -15,6 +14,5 @@ router.use('/image', imageRouter);
 router.use('/login', loginRouter);
 router.use('/user', userRouter);
 router.use('/comment', commentRouter);
-router.use('/search', searchRouter);
 
 export default router;
