+++ server/src/services/feed.service.ts
@@ -1,68 +1,42 @@
 import FeedModel, { IFeed, Iauthor } from '../models/feed.model';
-import UserModel, { IUser } from '../models/user.model';
+import { IUser } from '../models/user.model';
 
-const create = async (params: IFeed): Promise<boolean | undefined> => {
-  try {
-    const feed = new FeedModel(params);
-    const createResult = await feed.createFeed();
-
-    const { userId } = params.author;
-    const user = new UserModel({ userId });
-    await user.addFeedCount();
-
-    return createResult;
-  } catch (err) {
-    return undefined;
-  }
+const create = async (params: IFeed): Promise<boolean> => {
+  const feed = new FeedModel(params);
+  const result = await feed.createFeed();
+  return result;
 };
 
-const explore = async (lastFeedId: string): Promise<IFeed[] | undefined> => {
+const explore = async (): Promise<IFeed[]> => {
   const feed = new FeedModel();
-  try {
-    const result = await feed.exploreFeed(lastFeedId);
-    return result;
-  } catch (err) {
-    return undefined;
-  }
+  const result = feed.exploreFeed();
+  return result;
 };
 
-const following = async (
-  params: IUser,
-  lastFeedId?: string,
-): Promise<IFeed[] | undefined> => {
+const following = async (params: IUser): Promise<IFeed[]> => {
   const { follow: follows } = params;
-
   const userFollow: string[] = [];
   follows?.map((follow) => {
     if (follow.userId) {
       userFollow.push(follow.userId.toString());
     }
-    return follow;
   });
   const feed = new FeedModel();
-  try {
-    const result = await feed.followingFeed(userFollow, lastFeedId);
-    return result;
-  } catch (err) {
-    return undefined;
-  }
+  const result = feed.followingFeed(userFollow);
+  return result;
 };
 
 const like = async (
   author: Iauthor,
   feedId: string,
   status: number,
-): Promise<boolean | undefined> => {
+): Promise<boolean> => {
   const feed = new FeedModel();
-  try {
-    const result =
-      status === 1
-        ? await feed.addLike(author, feedId)
-        : await feed.deleteLike(author, feedId);
-    return result;
-  } catch (err) {
-    return undefined;
-  }
+  const result =
+    status === 1
+      ? await feed.addLike(author, feedId)
+      : await feed.deleteLike(author, feedId);
+  return result;
 };
 
 export { create, explore, following, like };
